[store] refactor: merge Meta into StateMachine
Meta is the user-data part in StateMachine.
The past I was stupid separating them.

Just moving codes, nothing changed.

24 LOC reduced!
