+++ fusestore/store/src/meta_service/mod.rs
@@ -2,16 +2,12 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-pub mod meta;
 pub mod meta_service_impl;
 pub mod placement;
 pub mod raftmeta;
 pub mod state_machine;
 
 pub use async_raft::NodeId;
-pub use meta::Meta;
-pub use meta::Node;
-pub use meta::Slot;
 pub use meta_service_impl::MetaServiceImpl;
 pub use placement::Placement;
 pub use raftmeta::ClientRequest;
@@ -22,7 +18,9 @@ pub use raftmeta::MetaNode;
 pub use raftmeta::Network;
 pub use raftmeta::RaftTxId;
 pub use raftmeta::ShutdownError;
-pub use state_machine::MemStoreStateMachine;
+pub use state_machine::Node;
+pub use state_machine::Slot;
+pub use state_machine::StateMachine;
 
 pub use crate::protobuf::meta_service_client::MetaServiceClient;
 pub use crate::protobuf::meta_service_server::MetaService;
@@ -33,8 +31,6 @@ pub use crate::protobuf::RaftMes;
 
 #[cfg(test)]
 mod meta_service_impl_test;
-#[cfg(test)]
-mod meta_test;
 #[cfg(test)]
 mod placement_test;
 #[cfg(test)]
